modules/location-service/requirements.txt
grpcio
grpcio-tools
python-dotenv
GeoAlchemy2